+++ tests/vfs.lua
@@ -1,10 +1,10 @@
 --[[
-	VFS Test
+	Graphene VFS Test
 ]]
 
-local N = require("namespace")
-local core = N:GetNamespaceCore()
-local vfs = core.FS:GetProvider("vfs")
+local Graphene = require("graphene")
+local G = Graphene:GetGrapheneCore()
+local vfs = G.FS:GetProvider("vfs")
 
 -- Basic write
 vfs:AddDirectory("Test")
@@ -15,13 +15,13 @@ vfs:AddDirectory("Test.Embedded")
 vfs:AddFile("Test.Embedded.X", [[ return {Name = "X"} ]])
 vfs:AddFile("Test.Embedded.Y", [[ return {Name = "Y", Friend = (...).X} ]])
 
-core:AddRebase("^Test%.Embedded%.", N.Test.Embedded)
+G:AddRebase("^Test%.Embedded%.", Graphene.Test.Embedded)
 
 -- FullyLoad capability
-local Test = N.Test
+local Test = Graphene.Test
 Test:FullyLoad()
 
-assert(N.HelloFlag, "FullyLoad failed to load Test.Hello")
+assert(Graphene.HelloFlag, "FullyLoad failed to load Test.Hello")
 
 assert(Test, "Could not add directory 'Test'")
 assert(Test.Name == "Test", "_ was not loaded for 'Test'")
@@ -30,3 +30,5 @@ assert(Test.Hello.Name == "Hello", "Hello did not load properly.")
 
 assert(Test.Embedded.X.Name == "X", "Rebased library did not load properly.")
 assert(Test.Embedded.Y.Friend.Name == "X", "Rebased library was not rebased correctly.")
+
+print("VFS tested successfully.")